offload ebpf - tether_stats_map - increase from 16 to 1000 entries
Use IFACE_STATS_MAP_SIZE (1000) not 16 as size of map.

Long term, it would be better to clean stuff up instead of letting
old interface stick around in the map, but in the mean time we should
match the size of the pre-existing iface_index_name_map since we require
that map to be functional anyway for statistics gathering to work.

Test: build, atest
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I17006664a7680d389281a3ce8b741f4380616298