netd_aidl_interface is marked as available to platform
The lib is depended on by services.net which is part of the platform.
With b/153073816, we will be checking the availability to the platform,
i.e. platform-available module depending on platform-unavailable module
causes a build error.

Prepare for the change by correctly marking that this interface lib is
available to the platform.

Bug: 153073816
Test: m

Change-Id: I971c1508b049524f3cc54b1060cf130ff66e02f3