Mark netd_aidl_interfaces-platform-java as available to platform
The lib is depended on by services.net that is part of the platform.

With b/153073816, we will be checking the availability to the platform,
i.e. platform-available module depending on platform-unavailable module
causes a build error.

Prepare for the change by correctly marking that this interface lib is
available to the platform.

Bug: 153073816
Test: m
Change-Id: I0bfbf822c4cd1c82408929d309938abfa147abd2