Add tap support to TunInterface.
This allows callers to create a tap interface instead of a tun.
Also specify IFF_NO_PI in order to make it easier for users of
this class to read data from the tun. No existing users actually
check the data, so this is safe.

Bug: 150736748
Test: atest netd_integration_test
Change-Id: Ic2e46bb864978b969ae5c1b8c2bdc5379d298f5a
Merged-In: Ic2e46bb864978b969ae5c1b8c2bdc5379d298f5a