Revert hardcoding of cgroup path.
Hardcoding the path in the test is a problem because the path
used by the device code changed in R QPR1.

This means that the same test code will not pass on both R
builds and R QPR builds. Currently this test is blocking Pixel
QPR build qualification, because aosp/1407127 has not been
cherry-picked to android11-tests-dev.

That cherry-pick is under review as aosp/1479484 to unblock
Pixel. However, if we just merge that cherry-pick and then do
nothing else, then when VTS 11.0r2 is released, all devices
trying to pass VTS on an R build will fail.

Revert the hardcoding. That seems more correct than attempting to
check for the two possible paths one after the other, and it also
makes any test failures much easier to debug.

The harcoding was introduced to fix ASAN-related flakiness on
cuttlefish, but it seems like the fix for that problem should not
affect the test semantics. Additionally, this test seems quite
stable on device (passes 1000 times on a row), so the flakiness
might be fixed.

This reverts commit 6bacd4fa63f809b113c2a277ae2c68ff14fee17c.

Bug: 171770637
Test: BpfBasicTest passed 1000 times in a row on aosp_crosshatch-eng
Change-Id: If47f654599e1d29f3bb91b2795559dac53ebe0c4