Disable IdletimerAddRemoveInterface test on >=5.10
Temporarily disable the IdletimerAddRemoveInterface test on >=5.10. The
test is not broken -- it is finding a real problem -- but we don't have
a fix for the kernel ready immediately, which blocks upgrading the
kernel for CF, GF and other devices to 5.10 (GKI 2.0).

When the kernel is fixed, we can revert this change and restore the
test.

Bug: 175745224
Change-Id: I084e93c5f7ce8c5f737ef997075c5ed5d86e930c