Remove netd_aidl_interfaces-platform-java
netd_aidl_interfaces-platform-java is only used by system service,
and now it's replaced by netd-client, so remove it.

Add tethering in apex_available to fulfil the dependency because
tethering links to netd-client which also links to
netd_event_listener_interface, also netd_event_listener_interface
is already maintained to be used in APEXes.
Set the min_sdk_version of netd_event_listener_interface to
29 to support the module which links to this library and which may
be installed on the Android Q.

Bug: 174837473
Test: make; atest FrameworksNetTests
Change-Id: I4ab2006ec73fd00a636c73946269f84193198511