OffloadUtils - swap clat and tether tc priorities (tether first)
It effectively doesn't currently matter what the ordering is because
the CLAT IPv6 destination IP will never be placed in the ingress tethering
offload map.

However in the future it makes sense for tethering (with built-in clat)
to be able to handle specific flows, before the generic clat translation
kicks in.

Additionally XDP happens before TC, so XDP tether will be before
TC tether, before TC clat offload.  This way the two near identical
tethering offloads will be next to each other.

Test: atest, TreeHugger
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: I7cac014b29d166dca9a8fd7739da86a7962d2190