Fix NetdBinderTest#InitialNetworksExist.
This test creates a socket with an implicit mark for an invalid
network (dummy or local), calls connect() on it, and expects
ENETUNREACH. Unfortunately this doesn't work because the connect()
overwrites the implicit mark with the default network mark, and
if the default network has IPv6 connectivity, the connect will
succeed.

Fix this by using an explicitly selected mark instead.

Bug: 179445508
Test: netd_integration_test passes with fix, fails without
Change-Id: Ib7848454e17d22cc58dfd096630d9d7ce503de04