Set min_sdk_version for netd_event_listener_interface-lateststable-java
The companion netd_aidl_interface-lateststable-java interface
already has a min_sdk version.

The merged-in tag is from aosp/1535680, which is only loosely
related (it's the CL that introduced this target). It's here
because this CL (and in general none of this topic) should reach
any downstream branches. Those branches will get this change when
it is cherry-picked to AOSP. That has to happen in a separate CL
anyway, because system/netd does not automerge to mainline-prod.

Test: m
Bug: 167645754
Change-Id: I4213db6693e981fe9d14bb1bf29b420fcf6737a4