Set min_sdk_version for netd_event_listener_interface-lateststable-java
The companion netd_aidl_interface-lateststable-java interface
already has a min_sdk version.

This is needed to make the tethering module updatable, which is
necessary to release the module. This is currently not the case
in AOSP due to ongoing development, but it has to be the case in
mainline-prod. This change is already present in mainline-prod,
and there is no harm to have it in AOSP as well.

Test: m
Bug: 167645754
Change-Id: I4213db6693e981fe9d14bb1bf29b420fcf6737a4