Change OemNetdListener::getListener to create a sp<OemNetdListener>
The test is broken because libbinder use stronger refbase semantics.
OemNetdListener::getListener() used to pass the raw pointer of the
instance to get sp<IBinder>, which breaks the stronger refbase semantics
(see aosp/1660499).

OemNetdListener::getListener() is now changed to create a
sp<OemNetdListener> to own the instance so that the code won't abort
in RefBase::incStrongRequireStrong().

Bug: 185191546
Test: atest netd_integration_test
Change-Id: I76920098142e86b0c035181d11d07c9cc639f065