Deflake NetlinkListenerTest#TestSkDestroyError
Two kinds of failure were observed,

Case 1: SkDestroyListen thread was unable to digest all events before
the test main thread timed out.
Case 2: All sockets are closed without ENOBUFS.

For case 1, extend the waiting time a bit. For case 2, have retries on
it. The fail rate is about 1 ~ 2% before this patch. The failure is not
observed after this patch in 1000 iterations.

Bug: 184991666
Bug: 174166169
Test: atest --rerun-until-failure=1000 NetlinkListenerTest
Change-Id: I72a7a41e415b704455c5165f9c9eb210762846e7