Keep using the latest stable version
... until the ToT version diverses from the latest stable version.

aosp/1659937 has created the version 6 of netd_aidl_interface. In doing
so, the references to the version V6 (formaly was unstable version) were
switched to the version V7. e.g.

shared_libs: {
- "netd_aidl_interface-V6-cpp",
+ "netd_aidl_interface-V7-cpp",
}

However, this is not only unnecessary (because at this moment V6 == V7
in terms of API signature), but also preventing the API freeze process
because the use an unstable version like V7 is strictly prohibited after
the API is frozen.

The expected development workflow is as follows:

* Now, V10 is the latest stable version. The team hasn't yet started
the development of V11.
* So, there is no reference to V11 at this moment. Every references are
either towards V10 or below.
* The team then starts the development of V11. This happens only in the
development branches like aosp/master, internal/master, but not in
<release>-dev branches.
* Some modules are starting to reference V11 if they need to use the
added APIs.
* V11 is frozen as the new stable version. The story goes on...

Bug: 188713899
Test: m
Change-Id: I13b1dc639723795113e8a6fa6427fccdcbe0c517