Never send packets with a source of ::1 on the wire.
Doing so is obviously invalid, and certain carriers will tear
down the connection if such packets are sent on their network.

This is done by adding an ip6tables rule to fw_OUTPUT that drops
all packets with a non-lo egress interface and a source of ::1.

Test: boot device, "adb root && adb shell ip6tables-save | egrep fw_OUTPUT"
Bug: 190368103
Bug: 198896920
Bug: 203096965
Signed-off-by: Maciej Żenczykowski <maze@google.com>
Change-Id: Ifb272d48705ba756ccd7bac806e4dc2dd7488cd5
Merged-In: Ifb272d48705ba756ccd7bac806e4dc2dd7488cd5